requirements.txt
django>=4.2.0
python-dotenv
psycopg2-binary
requests
werkzeug
jsonschema
openai
gunicorn
pillow
python-decouple
django-rest-framework